Specification fix: ensure that the initial thread can be referenced
With the previous specification, the initially active thread (as created
by starting a program) could not be referenced in transitions without
violating the specification (only threads specified in createThread
could be referenced).

Includes a revert of 1a839715dc2a0eb656c076957306ea3ee51f79b3, as the
check can now proceed as the ambiguity is resolved.