Simplify validation instructions for CPAchecker
For people who are users of the GraphML format the instructions on
how to configure CPAchecker to do a witness validation with a
non-standard analysis should not be listed here. If at all,
they should be listed in the CPAchecker documentation. The plan
is to eventually move these instructions to:
https://sosy-lab.gitlab.io/research/tutorials/CPAchecker/Witnesses.html
along with some additional information on how to do the same with
correctness witnesses.